refactor: rewrite api/v1beta1 → api/operator/v1beta1 import paths
Updates all Go import paths from api/v1beta1 to api/operator/v1beta1.
Package name stays v1beta1 — no code changes at usage sites.
No behavior changes.
