+++ internal/controller/file_controller.go
@@ -28,7 +28,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	opgmodels "github.com/neonephos-katalis/opg-ewbi-operator/api/ewbi/models"
-	"github.com/neonephos-katalis/opg-ewbi-operator/api/v1beta1"
+	"github.com/neonephos-katalis/opg-ewbi-operator/api/operator/v1beta1"
 	"github.com/neonephos-katalis/opg-ewbi-operator/internal/indexer"
 	"github.com/neonephos-katalis/opg-ewbi-operator/internal/multipart"
 	"github.com/neonephos-katalis/opg-ewbi-operator/internal/opg"
