+++ pkg/metastore/file.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/neonephos-katalis/opg-ewbi-operator/api/ewbi/models"
 	camara "github.com/neonephos-katalis/opg-ewbi-operator/api/ewbi/server"
-	opgv1beta1 "github.com/neonephos-katalis/opg-ewbi-operator/api/v1beta1"
+	opgv1beta1 "github.com/neonephos-katalis/opg-ewbi-operator/api/operator/v1beta1"
 )
 
 type File struct {
