+++ pkg/metastore/partneravailabilityzone.go
@@ -1,7 +1,7 @@
 package metastore
 
 import (
-	opgv1beta1 "github.com/neonephos-katalis/opg-ewbi-operator/api/v1beta1"
+	opgv1beta1 "github.com/neonephos-katalis/opg-ewbi-operator/api/operator/v1beta1"
 
 	"github.com/neonephos-katalis/opg-ewbi-operator/api/ewbi/models"
 )
